python/requirements.txt
python-dotenv==1.0.0
firebase-admin==6.2.0
